chore(chain): move queries to their own file
